now get the same result as fortran code

+++ src/c_tests/test_refvap.c
@@ -6,13 +6,12 @@
 
 int main(int argc, char *argv[]){
   
-  double nu = 22000000000.0; // 22 GHz
-  //double nu = 2.20000010E+10;
-  double t = 293.0; // Roughly 20 Celcius
-  double pvap = 2317.0;
-  double pdry = 101325.0 - pvap;
+  float nu = 22000000000.0; // 22 GHz
+  float t = 293.0; // Roughly 20 Celcius
+  float pvap = 2317.0;
+  float pdry = 101325.0 - pvap;
 
-  double complex nvap = refvap(nu, t, pdry, pvap);
+  float complex nvap = refvap(nu, t, pdry, pvap);
   fprintf(stdout, "%.8E\n", nu);
   fprintf(stdout, "%f\n", t);
   fprintf(stdout, "%f\n", pdry);
